Backend/requirements.txt
# # ─── Django core ──────────────────────────────────────────────────
# django==5.2.5
# djangorestframework==3.16.1
# djangorestframework_simplejwt==5.5.0
# django-cors-headers==4.7.0
# django-filter==25.1.0
# cryptography==45.0.6
# python-dotenv==1.1.1

# # ─── Database ─────────────────────────────────────────────────────
# psycopg2-binary==2.9.10

# # ─── Image handling ───────────────────────────────────────────────
# pillow==11.3.0

# # ─── Celery + RabbitMQ ────────────────────────────────────────────
# celery==5.5.3
# pika==1.3.2
# django-celery-results         # ← no pin: lets pip pick Django 5.2-compatible version
# django-celery-beat            # ← no pin: 2.7.0 blocks Django>=5.2, unpinned fixes it
# amqp==5.3.1
# pytz==2025.2
# python-dateutil==2.9.0

# # ─── Cache ────────────────────────────────────────────────────────
# django-redis==6.0.0

# # ─── API docs ─────────────────────────────────────────────────────
# drf-spectacular
# drf-spectacular-sidecar

# # ─── CKEditor ─────────────────────────────────────────────────────
# django-ckeditor-5==0.2.18

# # ─── Production server ────────────────────────────────────────────
# gunicorn==23.0.0

# # ═════════════════════════════════════════════════════════════════
# #  NEW — AI Resume Screener (RAG) packages
# # ═════════════════════════════════════════════════════════════════

# # LLM — Groq
# groq>=0.9.0
# tenacity>=8.3.0

# # LangChain + LangGraph
# langchain>=0.2.5
# langchain-groq>=0.1.6
# langgraph>=0.1.0

# # Web search
# tavily-python>=0.3.3

# # Embeddings — sentence-transformers (local, no API key)
# sentence-transformers>=3.0.0
# torch>=2.0.0

# # Vector store
# chromadb>=0.5.3

# # Resume file parsing
# pdfplumber>=0.11.0
# PyMuPDF>=1.24.0
# python-docx>=1.1.0





# ─── Django core ──────────────────────────────────────────────────
django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
django-filter==25.1.0
cryptography==45.0.6
python-dotenv==1.1.1

# ─── Database ─────────────────────────────────────────────────────
psycopg2-binary==2.9.10

# ─── Image handling ───────────────────────────────────────────────
pillow==11.3.0

# ─── Celery + RabbitMQ ────────────────────────────────────────────
celery==5.5.3
pika==1.3.2
django-celery-results>=2.6.0    # unpinned — lets pip find Django 5.2 compatible version
django-celery-beat>=2.7.0       # unpinned — lets pip find Django 5.2 compatible version
amqp==5.3.1
pytz==2025.2
python-dateutil==2.9.0

# ─── API docs ─────────────────────────────────────────────────────
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1

# ─── CKEditor ─────────────────────────────────────────────────────
django-ckeditor-5==0.2.18

# ─── Production server ────────────────────────────────────────────
gunicorn==23.0.0

# ═════════════════════════════════════════════════════════════════
#  AI Resume Screener (RAG) packages
# ═════════════════════════════════════════════════════════════════

# LLM — Groq
groq>=0.9.0
tenacity>=8.3.0

# LangChain + LangGraph
langchain>=0.2.5
langchain-groq>=0.1.6
langgraph>=0.1.0

# Web search
tavily-python>=0.3.3

# ── PyTorch CPU-only (~200 MB instead of ~2 GB CUDA build) ────────
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.3

# Resume file parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0
python-docx>=1.1.0